requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.42.0
paramiko>=3.5.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
pyyaml>=6.0.2
httpx>=0.28.0
python-dotenv>=1.0.1
